requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
future==0.17.1
idna==2.8
Markdown==3.0.1
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
pytz==2018.9
requests==2.21.0
stripe==2.22.0
urllib3==1.24.1
